requirements.txt
asgiref==3.5.2
tzdata
boto3==1.24.15
botocore==1.27.15
Django==3.2.9
django-cors-headers==3.13.0
django-storages==1.12.3
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
psycopg2-binary
django-jazzmin
django-shortcuts==1.6
djangorestframework-simplejwt
django-filter
requests
drf-yasg
django-ckeditor==6.2.0
requests_toolbelt
python-decouple
dataclasses-json
py3dns==3.2.1
validate-email-address==1 
drf-spectacular==0.24.2
drf-spectacular[sidecar]